parsers adapted to v.2 added default and mainQueue context to the proxy 'favorite' storing refactoring start [touch:143]{t:480}

+++ DrupalCon/DrupalCon/Model/Categories/DCEvent+DC.h
@@ -56,10 +56,10 @@ typedef NS_ENUM (int, DCEventType) {
 
 + (void)parseEventFromDictionaty:(NSDictionary*)eventDict toObject:(DCEvent*)object forDate:(NSDate*)date;
 
-- (void)addTypeForID:(NSInteger)typeID;
+- (void)addTypeForID:(int)typeID;
 - (void)addSpeakersForIds:(NSArray*)speakerIds;
-- (void)addLevelForID:(NSInteger)levelID;
-- (void)addTrackForId:(NSInteger)trackId;
+- (void)addLevelForID:(int)levelID;
+- (void)addTrackForId:(int)trackId;
 
 - (UIImage*)imageForEvent;
